update: register controller for dynamic client registration and signup for user
